# adapter/in/http/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Admin A username
type Admin struct {
	User *string `json:"user,omitempty"`
}

// AdminsArray Array of tags
type AdminsArray = []Admin

// AdminsArrayType Payload for admins
type AdminsArrayType struct {
	Admins *[]Admin `json:"admins,omitempty"`
}

// AuthenticationRequest Authenticate
type AuthenticationRequest struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// CompanyPayload Payload for creating an organisation
type CompanyPayload struct {
	Admins      *[]interface{} `json:"admins,omitempty"`
	Description *string        `json:"description,omitempty"`
	Fqdn        *string        `json:"fqdn,omitempty"`
	Name        string         `json:"name"`
}

// CreateOrganisationRequest defines model for CreateOrganisationRequest.
type CreateOrganisationRequest struct {
	Admins      *[]interface{} `json:"admins,omitempty"`
	Description *string        `json:"description,omitempty"`
	Fqdn        *string        `json:"fqdn,omitempty"`
	Name        string         `json:"name"`
}

// Dummy Dummy response
type Dummy struct {
	Message string `json:"message"`
}

// Error Error response
type Error struct {
	Message string `json:"message"`
}

// GetAllOrganisations Array of Organisations
type GetAllOrganisations = []Organisation

// ObjectId ObjectId
type ObjectId struct {
	Id *openapi_types.UUID `json:"id,omitempty"`
}

// Organisation defines model for Organisation.
type Organisation struct {
	Admins      *[]interface{}      `json:"admins,omitempty"`
	Description *string             `json:"description,omitempty"`
	Fqdn        *string             `json:"fqdn,omitempty"`
	Id          *openapi_types.UUID `json:"id,omitempty"`
	Name        string              `json:"name"`
}

// GetAllOrganisationsParams defines parameters for GetAllOrganisations.
type GetAllOrganisationsParams struct {
	// Limit Size of the page, maximum is 100, default is 25
	Limit *int32 `form:"limit,omitempty" json:"limit,omitempty"`

	// Page The page to return
	Page *int32 `form:"page,omitempty" json:"page,omitempty"`
}

// AuthenticateLoginJSONRequestBody defines body for AuthenticateLogin for application/json ContentType.
type AuthenticateLoginJSONRequestBody = AuthenticationRequest

// CreateOrganisationJSONRequestBody defines body for CreateOrganisation for application/json ContentType.
type CreateOrganisationJSONRequestBody = CreateOrganisationRequest